requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-annoying==0.10.4
idna==2.8
oauthlib==3.0.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tweepy==3.7.0
urllib3==1.25.3
